backend/pyproject.toml
[project]
name = "pbam"
version = "0.1.0"
description = "Private Banking & Analytic Management — backend"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    # Auth
    "python-jose[cryptography]>=3.3.0",
    "argon2-cffi>=23.1.0",
    # Validation / settings
    "pydantic[email]>=2.10.0",
    "pydantic-settings>=2.7.0",
    # OCR
    "easyocr>=1.7.2",
    "pdf2image>=1.17.0",
    "pillow>=11.0.0",
    # File handling
    "python-multipart>=0.0.20",
    # Utilities
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pbam"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["B008"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
